+++ src/api/scope/lib/latest-versions.js
@@ -1,0 +1,6 @@
+// @flow
+import { loadScope } from '../../../scope';
+
+export default function latestVersions(path: string, componentIds: BitId[]): Promise<any> {
+  return loadScope(path).then(scope => scope.latestVersions(componentIds));
+}