+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add `--outdated` flag to `bit list` command to show the remote versions of components
+
 ## [0.10.9] - 2017-10-18
 
 - rename `bit commit` to `bit tag`